requirements.txt
aiohttp==3.6.2
aiojobs==0.2.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
djangorestframework==3.11.0
idna==2.8
multidict==4.7.4
pyaes==1.6.1
Pyrogram==0.16.0
PySocks==1.7.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
TgCrypto==1.2.0
urllib3==1.25.7
yarl==1.4.2
